packages/core/pyproject.toml
[project]
name = "terrascope-core"
version = "0.1.0"
description = "Terrascope shared core: models, schemas, services, config"
requires-python = ">=3.14"
dependencies = [
    "pydantic-settings>=2.13",
    "sqlmodel>=0.0.24",
    "geoalchemy2>=0.17",
    "asyncpg>=0.31",
    "rasterio>=1.4",
    "geopandas>=1.0",
    "shapely>=2.0",
    "pyogrio>=0.10",
    "pyproj>=3.7",
    "torch>=2.6",
    "torchvision>=0.21",
    "ultralytics>=8.3",
    "sahi>=0.11",
    "transformers>=4.40",
    "Pillow>=10",
    "pystac-client>=0.8",
    "matplotlib>=3.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core"]
